Adds fixed random seed of 42.
Using a fixed random seed ensures that when changes are made to the
top-level code that don't affect the output, re-running the notebook
doesn't cause a change in the binary outputs. This makes version control
easier.

It also makes it easier for end users to compare their results with our
results.